+++ commons/src/main/java/de/gematik/refv/commons/configuration/DependencyListsWrapper.java
@@ -1,19 +1,18 @@
 /*
- * Copyright (c) 2024 gematik GmbH
- * 
- * Licensed under the Apache License, Version 2.0 (the License);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an 'AS IS' BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+Copyright (c) 2022-2024 gematik GmbH
 
+Licensed under the Apache License, Version 2.0 (the License);
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an 'AS IS' BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 package de.gematik.refv.commons.configuration;
 
 import lombok.AllArgsConstructor;
@@ -21,6 +20,7 @@ import lombok.AllArgsConstructor;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @AllArgsConstructor
 public class DependencyListsWrapper {
@@ -31,17 +31,13 @@ public class DependencyListsWrapper {
     }
 
     public Optional<DependencyList> getDependencyListValidAt(LocalDate resourceCreationDate) {
-        for (DependencyList list :
-                dependencyLists) {
-            var dateFrom = LocalDate.parse(list.getValidFrom());
-            var dateTill = list.getValidTill() == null ? LocalDate.MAX : LocalDate.parse(list.getValidTill());
-            if(
-                    (list.getValidFrom() != null && (dateFrom.isBefore(resourceCreationDate) || dateFrom.isEqual(resourceCreationDate))) &&
-                    (list.getValidTill() == null || dateTill.isAfter(resourceCreationDate) || dateTill.isEqual(resourceCreationDate))
-            )
-                    return Optional.of(list);
-        }
-        return Optional.empty();
+        var matchingDependencyLists = dependencyLists.stream().filter(dl -> dl.isValidFor(resourceCreationDate)).collect(Collectors.toList());
+        if(matchingDependencyLists.size() > 1)
+            throw new IllegalStateException(String.format("Multiple dependency lists are valid for the date %s",resourceCreationDate));
+        if(matchingDependencyLists.size() == 0)
+            return Optional.empty();
+
+        return Optional.of(matchingDependencyLists.get(0));
     }
 
     public DependencyList getLatestDependencyList() {